services/ocr/requirements.txt
numpy==1.26.4
opencv-python-headless==4.10.0.84
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
paddlepaddle==2.6.1
paddleocr==2.7.3
python-multipart==0.0.9
